Backend/models/Weapon_Detection/requirements.txt
ultralytics
opencv-python
torch
python-telegram-bot